Encodage des notes Move folder shared_kernel/service to /shared_kernel/domain/service
