+++ ddd/logic/encodage_des_notes/encodage/domain/service/rechercher_notes_etudiant.py
@@ -28,10 +28,9 @@ from base.models.enums.exam_enrollment_justification_type import JustificationTy
 from ddd.logic.encodage_des_notes.encodage.domain.model._note import NOTE_MANQUANTE
 from ddd.logic.encodage_des_notes.encodage.domain.model.gestionnaire_parcours import GestionnaireParcours
 from ddd.logic.encodage_des_notes.encodage.domain.model.note_etudiant import NoteEtudiant
-from ddd.logic.encodage_des_notes.encodage.domain.service.i_cohortes_du_gestionnaire import ICohortesDuGestionnaire
 from ddd.logic.encodage_des_notes.encodage.repository.note_etudiant import INoteEtudiantRepository
 from ddd.logic.encodage_des_notes.shared_kernel.dtos import NoteEtudiantDTO
-from ddd.logic.encodage_des_notes.shared_kernel.service.i_signaletique_etudiant import ISignaletiqueEtudiantTranslator
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service import ISignaletiqueEtudiantTranslator
 from osis_common.ddd import interface
