+++ ddd/logic/encodage_des_notes/encodage/use_case/read/rechercher_notes_service.py
@@ -31,9 +31,9 @@ from ddd.logic.encodage_des_notes.encodage.domain.service.i_cohortes_du_gestionn
 from ddd.logic.encodage_des_notes.encodage.domain.service.rechercher_notes_etudiant import RechercheNotesEtudiant
 from ddd.logic.encodage_des_notes.encodage.repository.note_etudiant import INoteEtudiantRepository
 from ddd.logic.encodage_des_notes.shared_kernel.dtos import NoteEtudiantDTO
-from ddd.logic.encodage_des_notes.shared_kernel.service.i_periode_encodage_notes import IPeriodeEncodageNotesTranslator
-from ddd.logic.encodage_des_notes.shared_kernel.service.i_signaletique_etudiant import ISignaletiqueEtudiantTranslator
-from ddd.logic.encodage_des_notes.shared_kernel.service.periode_encodage_ouverte import PeriodeEncodageOuverte
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service import IPeriodeEncodageNotesTranslator
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service import ISignaletiqueEtudiantTranslator
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service.periode_encodage_ouverte import PeriodeEncodageOuverte
 
 
 def rechercher_notes(
