+++ ddd/logic/encodage_des_notes/shared_kernel/domain/service/i_inscription_examen.py
