+++ ddd/logic/encodage_des_notes/shared_kernel/domain/service/i_periode_encodage_notes.py
