+++ ddd/logic/encodage_des_notes/shared_kernel/domain/service/i_signaletique_personne.py
