+++ ddd/logic/encodage_des_notes/shared_kernel/domain/service/periode_encodage_ouverte.py
@@ -23,21 +23,29 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-import abc
-from typing import Set, List
+import datetime
 
-from ddd.logic.encodage_des_notes.soumission.dtos import AttributionEnseignantDTO, SignaletiqueEtudiantDTO
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service.i_periode_encodage_notes import \
+    IPeriodeEncodageNotesTranslator
+from ddd.logic.encodage_des_notes.shared_kernel.validator.exceptions import PeriodeEncodageNotesFermeeException
 from osis_common.ddd import interface
 
 
-class ISignaletiqueEtudiantTranslator(interface.DomainService):
+class PeriodeEncodageOuverte(interface.DomainService):
 
     @classmethod
-    @abc.abstractmethod
-    def search(
+    def verifier(
             cls,
-            nomas: List[str],
-            nom: str = None,
-            prenom: str = None,
-    ) -> Set['SignaletiqueEtudiantDTO']:
-        raise NotImplementedError
+            periode_soumission_note_translator: 'IPeriodeEncodageNotesTranslator'
+    ) -> None:
+        periode = periode_soumission_note_translator.get()
+        if not periode:
+            raise PeriodeEncodageNotesFermeeException()
+
+        aujourdhui = datetime.date.today()
+        debut_periode = periode.debut_periode_soumission.to_date()
+        fin_periode = periode.fin_periode_soumission.to_date()
+        periode_est_ouverte = debut_periode <= aujourdhui <= fin_periode
+
+        if not periode_est_ouverte:
+            raise PeriodeEncodageNotesFermeeException()
