+++ ddd/logic/encodage_des_notes/soumission/domain/service/donnees_administratives.py
@@ -26,8 +26,8 @@
 import itertools
 from typing import List
 
-from ddd.logic.encodage_des_notes.shared_kernel.service.i_inscription_examen import IInscriptionExamenTranslator
-from ddd.logic.encodage_des_notes.shared_kernel.service.i_periode_encodage_notes import IPeriodeEncodageNotesTranslator
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service import IInscriptionExamenTranslator
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service import IPeriodeEncodageNotesTranslator
 from ddd.logic.encodage_des_notes.soumission.domain.service.i_contact_feuille_de_notes import \
     IAdresseFeuilleDeNotesTranslator
 from ddd.logic.encodage_des_notes.soumission.domain.service.i_deliberation import IDeliberationTranslator
