+++ ddd/logic/encodage_des_notes/soumission/use_case/read/search_donnees_administratives_feuille_de_notes_service.py
@@ -25,8 +25,8 @@
 ##############################################################################
 from typing import List
 
-from ddd.logic.encodage_des_notes.shared_kernel.service.i_inscription_examen import IInscriptionExamenTranslator
-from ddd.logic.encodage_des_notes.shared_kernel.service.i_periode_encodage_notes import IPeriodeEncodageNotesTranslator
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service import IInscriptionExamenTranslator
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service import IPeriodeEncodageNotesTranslator
 from ddd.logic.encodage_des_notes.soumission.commands import SearchAdressesFeuilleDeNotesCommand
 from ddd.logic.encodage_des_notes.soumission.domain.service.donnees_administratives import DonneesAdministratives
 from ddd.logic.encodage_des_notes.soumission.domain.service.i_contact_feuille_de_notes import \
