+++ infrastructure/encodage_de_notes/shared_kernel/service/in_memory/inscription_examen.py
@@ -25,7 +25,7 @@
 ##############################################################################
 from typing import Set
 
-from ddd.logic.encodage_des_notes.shared_kernel.service.i_inscription_examen import IInscriptionExamenTranslator
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service import IInscriptionExamenTranslator
 from ddd.logic.encodage_des_notes.soumission.dtos import InscriptionExamenDTO, DesinscriptionExamenDTO
 from ddd.logic.encodage_des_notes.shared_kernel.dtos import DateDTO
