+++ infrastructure/encodage_de_notes/shared_kernel/service/in_memory/signaletique_etudiant.py
@@ -26,7 +26,7 @@
 from typing import Set, List
 
 from base.models.enums.peps_type import PepsTypes
-from ddd.logic.encodage_des_notes.shared_kernel.service.i_signaletique_etudiant import \
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service import \
     ISignaletiqueEtudiantTranslator
 from ddd.logic.encodage_des_notes.soumission.dtos import SignaletiqueEtudiantDTO
 from ddd.logic.encodage_des_notes.shared_kernel.dtos import EtudiantPepsDTO
