+++ infrastructure/encodage_de_notes/shared_kernel/service/signaletique_etudiant.py
@@ -28,7 +28,7 @@ from typing import Set, List, Dict
 from django.db.models import F
 
 from base.models.student import Student
-from ddd.logic.encodage_des_notes.shared_kernel.service.i_signaletique_etudiant import \
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service import \
     ISignaletiqueEtudiantTranslator
 from ddd.logic.encodage_des_notes.soumission.dtos import SignaletiqueEtudiantDTO
 from ddd.logic.encodage_des_notes.shared_kernel.dtos import EtudiantPepsDTO
