+++ infrastructure/encodage_de_notes/soumission/domain/service/in_memory/signaletique_personne.py
@@ -25,7 +25,7 @@
 ##############################################################################
 from typing import Set
 
-from ddd.logic.encodage_des_notes.shared_kernel.service.i_signaletique_personne import ISignaletiquePersonneTranslator
+from ddd.logic.encodage_des_notes.shared_kernel.domain.service.i_signaletique_personne import ISignaletiquePersonneTranslator
 from ddd.logic.encodage_des_notes.shared_kernel.dtos import DetailContactDTO, AdresseDTO
